Branch↔workspace-id mapping moves to src/shared — it was never workspace-service logic
workspaceIdForBranch/branchForWorkspaceId are the pure naming convention
every module computes with; importing them from workspace.service dragged
the whole workspace module (and its fs/git machinery) into access,
tool-manuals, plugins, skills, diff and workflow. They now live in
src/shared/workspace-id.ts with no module imports at all. Clean break —
workspace.service no longer re-exports them; the package's public index
never surfaced them.
