+++ packages/core-backend/src/modules/access/synced-groups-committer.ts
@@ -1,7 +1,7 @@
 import path from 'node:path';
 import type { AuthUser, IWorkflowService } from '@bevel-software/platform-shared';
 import type { WorkspaceService } from '../workspace/workspace.service.js';
-import { workspaceIdForBranch } from '../workspace/workspace.service.js';
+import { workspaceIdForBranch } from '../../shared/workspace-id.js';
 import type { AccessControlService } from './access-control.service.js';
 import { SYNCED_GROUPS_YAML } from './group-files.js';
 import type { WorkflowEventBus } from '../workflow/event-bus.js';
