+++ packages/core-backend/src/modules/diff/__tests__/diff.routes.rejectPathsLocked.test.ts
@@ -4,7 +4,7 @@ import path from 'node:path';
 import os from 'node:os';
 import type { AuthUser, IWorkflowService } from '@bevel-software/platform-shared';
 import type { IDiffService } from '../diff.interface.js';
-import { workspaceIdForBranch } from '../../workspace/workspace.service.js';
+import { workspaceIdForBranch } from '../../../shared/workspace-id.js';
 import { WorkflowValidationError } from '../../workflow/workflow.errors.js';
 import { rejectPathsLocked } from '../diff.routes.js';
