+++ packages/core-backend/src/modules/plugins/__tests__/join-requests.service.test.ts
@@ -6,10 +6,8 @@ import {
   type ChangeRequest,
   type IWorkflowService,
 } from '@bevel-software/platform-shared';
-import {
-  workspaceIdForBranch,
-  type WorkspaceService,
-} from '../../workspace/workspace.service.js';
+import { type WorkspaceService } from '../../workspace/workspace.service.js';
+import { workspaceIdForBranch } from '../../../shared/workspace-id.js';
 import { JoinRequestsService } from '../join-requests.service.js';
 
 /**
