+++ packages/core-backend/src/modules/workflow/workflow.routes.ts
@@ -27,7 +27,7 @@ import type {
 } from '@bevel-software/platform-shared';
 import type { AuthService } from '../auth/auth.service.js';
 import type { WorkspaceService } from '../workspace/workspace.service.js';
-import { branchForWorkspaceId } from '../workspace/workspace.service.js';
+import { branchForWorkspaceId } from '../../shared/workspace-id.js';
 import type { WorkflowEventBus } from './event-bus.js';
 import { WorkflowDomainError } from './workflow.errors.js';
 import '../auth/auth.middleware.js'; // Express Request augmentation
