+++ packages/core-backend/src/modules/workspace/__tests__/workspace.service.test.ts
@@ -4,7 +4,8 @@ import { promisify } from 'node:util';
 import fs from 'node:fs/promises';
 import path from 'node:path';
 import os from 'node:os';
-import { WorkspaceService, workspaceIdForBranch } from '../workspace.service.js';
+import { WorkspaceService } from '../workspace.service.js';
+import { workspaceIdForBranch } from '../../../shared/workspace-id.js';
 
 const execFileAsync = promisify(execFile);
