+++ packages/core-backend/src/modules/workspace/workspace.tools.ts
@@ -16,7 +16,7 @@ import {
 import type { IRoutineWritePolicy } from './routine-write-policy.js';
 import type { ToolHandlerFactory } from '../tool-helpers/tool-handler.js';
 import { requireInternalSource, requireExternalSource } from '../tool-auth/tool-auth.middleware.js';
-import { workspaceIdForBranch } from './workspace.service.js';
+import { workspaceIdForBranch } from '../../shared/workspace-id.js';
 // Leaf-level shared primitive (same exception `workspace.service.ts` already
 // relies on) — not a workflow service, so this stays inside the module boundary.
 import { assertValidBranchName } from '../workflow/git/branch-name.js';
